refactor: move buttonsObj to module
