+++ .eslintrc
@@ -3,6 +3,7 @@
   "rules": {
     "no-console": "off",
     "no-undef": "off",
-    "space-before-function-paren": "off"
-  }
+    "space-before-function-paren": "off",
+    "import/extensions": "always"
+  } 
 }
