+++ index.html
@@ -10,6 +10,6 @@
 </head>
 <body>
   
-  <script src="script.js"></script>
+  <script src="script.js" type="module"></script>
 </body>
 </html>
